Split the `cargo::util::network` module into submodules
This is intended to help grow with more stuff.
